requirements.txt
Django>=4.1.5,<4.2
djangorestframework>=3.14,<3.15
drf-spectacular>=0.25.1,<0.26
psycopg2>=2.9.5,<3.0
Pillow>=9.4.0,<9.5
gunicorn>=20.1.0,<20.2